extractor/requirements.txt
elasticsearch
python-docx
pdfplumber